backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic-settings==2.10.1
redis==6.4.0
python-multipart==0.0.20
aiofiles==24.1.0
orjson==3.10.18
numpy==2.2.6
soundfile==0.13.1
qwen-tts==0.1.0